[RMV] test_extendable_pydantic_fastapi: move tests to extendable_fastapi
