+++ extendable_fastapi/__manifest__.py
@@ -15,6 +15,7 @@
     "demo": [],
     "external_dependencies": {
         "python": [
+            "pydantic>=2.0.0",
             "extendable-pydantic>=1.1.0",
         ],
     },
