+++ extendable_fastapi/tests/__init__.py
@@ -1,1 +1,2 @@
+from . import test_generic_extendable
 from . import test_strict_extendable_base_model
