--- setup/test_extendable_pydantic_fastapi/odoo/addons/test_extendable_pydantic_fastapi
@@ -1,1 +1,0 @@
-../../../../test_extendable_pydantic_fastapi
