--- setup/test_extendable_pydantic_fastapi/setup.py
@@ -1,6 +1,0 @@
-import setuptools
-
-setuptools.setup(
-    setup_requires=['setuptools-odoo'],
-    odoo_addon=True,
-)
