--- test_extendable_pydantic_fastapi/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import schemas
