--- test_extendable_pydantic_fastapi/i18n/test_extendable_pydantic_fastapi.pot
@@ -1,19 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* test_extendable_pydantic_fastapi
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: test_extendable_pydantic_fastapi
-#: model:ir.model,name:test_extendable_pydantic_fastapi.model_fastapi_endpoint
-msgid "FastAPI Endpoint"
-msgstr ""
