--- test_extendable_pydantic_fastapi/models/__init__.py
@@ -1,1 +1,0 @@
-from . import fastapi_endpoint
