--- test_extendable_pydantic_fastapi/readme/CONTRIBUTORS.rst
@@ -1,1 +1,0 @@
-* Laurent Mignon <laurent.mignon@acsone.eu> (ACSONE SA/NV)
