--- test_extendable_pydantic_fastapi/readme/DESCRIPTION.rst
@@ -1,2 +1,0 @@
-This module is a technical addon used to tests the integration between odoo,
-fastapi and extendable_pydantic.
