--- test_extendable_pydantic_fastapi/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_user
